requirements.txt
alembic (0.9.9)
bcrypt (3.1.4)
cffi (1.11.5)
click (6.7)
Flask (1.0.2)
Flask-Migrate (2.1.1)
Flask-SQLAlchemy (2.3.2)
itsdangerous (0.24)
Jinja2 (2.10)
Mako (1.0.7)
MarkupSafe (1.0)
pip (9.0.3)
psycopg2 (2.7.4)
psycopg2-binary (2.7.4)
pycparser (2.18)
PyJWT (1.6.1)
python-dateutil (2.7.3)
python-dotenv (0.8.2)
python-editor (1.0.3)
setuptools (39.0.1)
six (1.11.0)
SQLAlchemy (1.2.7)
Werkzeug (0.14.1)
